MSSQL_HELLFIRE_FRAMEWORK/requirements.txt
# requirements.txt
# MSSQL HELLFIRE FRAMEWORK - DEPENDÊNCIAS COMPLETAS
# Última atualização: 2024

# 🔥 CORE DEPENDENCIES
requests==2.31.0
urllib3==2.0.7
beautifulsoup4==4.12.2
lxml==4.9.3

# 🗄️ DATABASE DRIVERS  
pymssql==2.2.7                  # MSSQL
mysql-connector-python==8.1.0   # MySQL
psycopg2-binary==2.9.7          # PostgreSQL
pymongo==4.5.0                  # MongoDB
redis==5.0.1                    # Redis

# 🔐 CRYPTO & SECURITY
cryptography==41.0.4
pycryptodome==3.19.0
fernet==1.0.1
bcrypt==4.0.1

# 🌐 NETWORK & PROTOCOLS
paramiko==3.3.1                 # SSH
smbprotocol==1.10.1             # SMB
ldap3==2.9.1                    # LDAP
ftplib==0.8.0                   # FTP
dnspython==2.4.2                # DNS
scapy==2.5.0                    # Packet manipulation

# 🕵️ RECON & SCANNING
nmap==1.0.1                     # Python-nmap
python-whois==0.8.0
geoip2==4.7.0                   # GeoIP lookup
shodan==1.29.0                  # Shodan API
censys==2.2.9                   # Censys API

# 🚀 WEB & EXPLOITATION
selenium==4.15.2
webdriver-manager==4.0.1
pyppeteer==1.0.2
mechanicalsoup==1.3.0
httpie==3.2.2

# 📊 DATA PROCESSING
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2
xlrd==2.0.1

# 🎭 EVASION & SPOOFING
fake-useragent==1.4.0
user-agents==2.2.0
pyvirtualdisplay==3.0
python-dateutil==2.8.2

# 📝 REPORTING & OUTPUT
Jinja2==3.1.2
MarkupSafe==2.1.3
python-docx==1.1.0
xlsxwriter==3.1.9
reportlab==4.0.6
pdfkit==1.0.0

# 🔧 UTILITIES & HELPERS
colorama==0.4.6
tqdm==4.66.1
progress==1.6
python-magic==0.4.27
filelock==3.13.1

# 🧠 ASYNC & PERFORMANCE
aiohttp==3.8.6
asyncio==3.4.3
uvloop==0.19.0
httpx==0.25.2

# 🐍 ADVANCED PYTHON
typing-extensions==4.8.0
python-dotenv==1.0.0
python-decouple==3.8
configparser==6.0.0

# 📡 API INTEGRATIONS
python-telegram-bot==20.7
discord.py==2.3.2
slack-sdk==3.27.0
twilio==8.13.0

# 🗃️ DATA SERIALIZATION
pyyaml==6.0.1
xmltodict==0.13.0
msgpack==1.0.7
orjson==3.9.10

# 🔍 DEBUGGING & LOGGING
loguru==0.7.2
rich==13.7.0
icecream==2.1.3
debugpy==1.8.0

# 🛡️ ADDITIONAL SECURITY
pyopenssl==23.2.0
certifi==2023.11.17
oauthlib==3.2.2
requests-oauthlib==1.3.1

# 🧩 MISCELLANEOUS
pillow==10.1.0
qrcode==7.4.2
barcode==1.0.1
pyfiglet==0.8.post1

# 📦 OPTIONAL (DESCOMENTE SE PRECISAR)
# tensorflow==2.14.0              # Machine Learning
# torch==2.1.1                    # PyTorch
# scikit-learn==1.3.2             # ML algorithms
# selenium-stealth==1.0.6         # Stealth selenium
# pyinstaller==5.13.0             # Build executables